+++ lib/component.ex
@@ -13,7 +13,7 @@ defmodule PrimerLive.Component do
   # action_list
   # ------------------------------------------------------------------------------------
 
-  @doc section: :navigation
+  @doc section: :menus
 
   @doc ~S"""
   Generates a vertical list of interactive actions or options. It’s composed of items presented in a consistent single-column format, with room for icons, descriptions, side information, and other rich visuals.
@@ -361,7 +361,7 @@ defmodule PrimerLive.Component do
   # action_list_section_divider
   # ------------------------------------------------------------------------------------
 
-  @doc section: :navigation
+  @doc section: :menus
 
   @doc ~S"""
   Action list section divider for separating groups of content. See `action_list/1`.
@@ -541,7 +541,7 @@ defmodule PrimerLive.Component do
   # action_list_item
   # ------------------------------------------------------------------------------------
 
-  @doc section: :navigation
+  @doc section: :menus
 
   @doc ~S"""
   Action list item. See `action_list/1`.
@@ -898,7 +898,6 @@ defmodule PrimerLive.Component do
                 form={@form}
                 field={@field}
                 checked={@is_selected}
-                value={@checked_value}
                 class={@classes.leading_visual_single_select_checkmark}
               >
                 <:label>
@@ -4233,6 +4232,14 @@ defmodule PrimerLive.Component do
     doc:
       "Either a [Phoenix.HTML.Form](https://hexdocs.pm/phoenix_html/Phoenix.HTML.Form.html) or an atom."
 
+  attr :checked, :boolean, doc: "The state of the checkbox (when not using `form` and `field`)."
+
+  attr :hidden_input, :string,
+    default: "true",
+    doc: """
+    Controls if the component will generate a hidden input to submit the unchecked checkbox value or not. Defaults to "true". Uses `Phoenix.HTML.Form.hidden_input/3`.
+    """
+
   attr :is_multiple, :boolean,
     default: false,
     doc:
@@ -4449,6 +4456,9 @@ defmodule PrimerLive.Component do
         [
           [id: input_id],
           [name: input_name],
+          !is_nil(assigns[:checked]) && [checked: assigns[:checked]],
+          assigns.input_type === :checkbox &&
+            [hidden_input: assigns.hidden_input],
           input_class && [class: input_class],
           show_message? && [invalid: ""]
         ]
@@ -4644,6 +4654,9 @@ defmodule PrimerLive.Component do
       "Either a [Phoenix.HTML.Form](https://hexdocs.pm/phoenix_html/Phoenix.HTML.Form.html) or an atom."
 
   attr(:value, :string, default: nil, doc: "Input value.")
+
+  attr :checked, :boolean,
+    doc: "The state of the radio button (when not using `form` and `field`)."
 
   attr(:is_emphasised_label, :boolean, default: false, doc: "Adds emphasis to the label.")
 
@@ -6993,8 +7006,7 @@ defmodule PrimerLive.Component do
       }">Menu</:toggle>
       ```
 
-      See also: See: https://github.com/ArthurClemens/dialogic-js#prompttoggle
-      ```
+      See also: https://github.com/ArthurClemens/dialogic-js#prompttoggle
       """
     )