+++ test/components/checkbox_test.exs
@@ -10,7 +10,7 @@ defmodule PrimerLive.TestComponents.CheckboxTest do
     impl: Phoenix.HTML.FormData.Atom,
     id: "user",
     name: "user",
-    params: %{"available_for_hire" => ""},
+    params: %{"available_for_hire" => "true"},
     source: %Ecto.Changeset{
       action: :validate,
       changes: %{},
@@ -215,7 +215,7 @@ defmodule PrimerLive.TestComponents.CheckboxTest do
              <span class="FormControl-checkbox-wrap container-x my-checkbox">
              <span phx-feedback-for="user[available_for_hire]" class="pl-valid"></span>
              <input name="user[available_for_hire]" type="hidden" value="false" />
-             <input class="form-checkbox-details-trigger FormControl-checkbox input-x" id="user[available_for_hire]" name="user[available_for_hire]" type="checkbox" value="true" />
+             <input checked class="form-checkbox-details-trigger FormControl-checkbox input-x" id="user[available_for_hire]" name="user[available_for_hire]" type="checkbox" value="true" />
              <span class="FormControl-checkbox-labelWrap label_container-x"><label class="FormControl-label label-x my-label" aria-live="polite" for="user[available_for_hire]">Some label</label>
              <p class="FormControl-caption hint-x my-hint">Some hint</p><span class="form-checkbox-details text-normal disclosure-x my-disclosure">Some hint</span>
              </span>